Modifiquei os arquivos config/custom-express, controllers/pagamentos, files/pagamentos. E criei os arquivos controllers/correios, controllers/uploads, files/coracao.png, files/dados-entrega.json, servicos/clienteCartoes, servicos/correios/correiosSOAPClient.js, util/coracao.png, util/fileReader, util/streamFileReader

+++ controllers/pagamentos.js
@@ -56,8 +56,8 @@ module.exports = function(app){
     });
     app.post('/pagamentos/pagamento', function(req, res){
 
-        req.assert('forma_de_pagamento', 'Forma de pagamento é obrigatório').notEmpty();
-        req.assert('valor', 'Valor é obrigatório e deve ser um decimal').notEmpty().isFloat();
+        req.assert('pagamento.forma_de_pagamento', 'Forma de pagamento é obrigatório').notEmpty();
+        req.assert('pagamento.valor', 'Valor é obrigatório e deve ser um decimal').notEmpty().isFloat();
 
         var erros = req.validationErrors();
         if(erros){
@@ -67,7 +67,7 @@ module.exports = function(app){
         }
       
 
-        var pagamento = req.body;
+        var pagamento = req.body["pagamento"];
 
         console.log('processando uma requisição de um novo pagamento');
 
@@ -86,26 +86,67 @@ module.exports = function(app){
                 pagamento.id = resultado.insertId;
                 
                 console.log('pagamento criado');  
-                res.location('/pagamentos/pagamento/' + pagamento.id);   
-                
-                var response = {
-                    dados_do_pagamento: pagamento,
-                    links: [
-                        {
-                            href: "http://localhost:80/pagamentos/pagamento/" + pagamento.id, 
-                            rel: "confirmar",
-                            method: "PUT"
-                        },
-                        {
-                            href: "http://localhost:80/pagamentos/pagamento/" + pagamento.id,
-                            rel: "cancelar",
-                            method: "DELETE"
-                        }
-                    ]
 
-                };
-                
-                res.status(201).json(response);
+                if(pagamento.forma_de_pagamento == 'cartao'){
+                    var cartao = req.body['cartao'];
+                    console.log(cartao);  
+
+                    var clienteCartoes = new app.servicos.clienteCartoes();
+                    clienteCartoes.autoriza(cartao, function(exception, request, response, retorno){
+                        if(exception){
+                            console.log(exception['message']);
+                            res.status(400).send(exception['message']);  
+                            return;
+                        }
+                        console.log(retorno);
+
+                        res.location('/pagamentos/pagamento' + pagamento.id);
+
+                        var response = {
+                            dados_do_pagamento: pagamento,
+                            cartao: retorno,
+                            links: [
+                                {
+                                    href:"http://localhost:80/pagamentos/pagamento" + pagamento.id,
+                                    rel: "confirmar",
+                                    method:"PUT"
+                                },
+                                {
+                                    href:"http://localhost:80/pagamentos/pagamento" + pagamento.id,
+                                    rel: "cancelar",
+                                    method:"DELETE"
+                                }
+                            ]
+                        }
+                        res.status(201).json(response);  
+                        return;
+
+                    });
+                   
+                  
+                } else {
+
+                    res.location('/pagamentos/pagamento/' + pagamento.id);   
+                    
+                    var response = {
+                        dados_do_pagamento: pagamento,
+                        links: [
+                            {
+                                href: "http://localhost:80/pagamentos/pagamento/" + pagamento.id, 
+                                rel: "confirmar",
+                                method: "PUT"
+                            },
+                            {
+                                href: "http://localhost:80/pagamentos/pagamento/" + pagamento.id,
+                                rel: "cancelar",
+                                method: "DELETE"
+                            }
+                        ]
+
+                    };
+                    
+                    res.status(201).json(response);
+                }
             }    
         });
